+++ bin/ci/phpunit.php
@@ -5,35 +5,35 @@ declare(strict_types=1);
 use Steevanb\ParallelProcess\{
     Console\Application\ParallelProcessesApplication,
     Process\Process,
-    Process\ProcessInterfaceArray
+    Process\ProcessInterfaceCollection
 };
-use Steevanb\PhpTypedArray\ScalarArray\StringArray;
+use Steevanb\PhpCollection\ScalarCollection\StringCollection;
 use Symfony\Component\Console\Input\ArgvInput;
 
 require dirname(__DIR__, 2) . '/vendor/autoload.php';
 
-function getAvailableSymfonyVersions(): StringArray
+function getAvailableSymfonyVersions(): StringCollection
 {
-    $return = new StringArray();
-    foreach (new StringArray(['6.1', '6.2', '6.3']) as $symfonyVersion) {
-        $return[] = $symfonyVersion;
+    $return = new StringCollection();
+    foreach (new StringCollection(['6.1', '6.2', '6.3']) as $symfonyVersion) {
+        $return->add($symfonyVersion);
     }
 
     return $return;
 }
 
-function createPhpunitProcesses(string $phpVersion = null, string $symfonyVersion = null): ProcessInterfaceArray
+function createPhpunitProcesses(string $phpVersion = null, string $symfonyVersion = null): ProcessInterfaceCollection
 {
-    $phpVersions = new StringArray(is_string($phpVersion) ? [$phpVersion] : ['8.1', '8.2']);
+    $phpVersions = new StringCollection(is_string($phpVersion) ? [$phpVersion] : ['8.1', '8.2']);
 
-    $return = new ProcessInterfaceArray();
+    $return = new ProcessInterfaceCollection();
     foreach ($phpVersions as $loopPhpVersion) {
         $symfonyVersions = is_string($symfonyVersion)
             ? [$symfonyVersion]
             : getAvailableSymfonyVersions()->toArray();
 
         foreach ($symfonyVersions as $loopSymfonyVersion) {
-            $return[] = createPhpunitProcess($loopPhpVersion, $loopSymfonyVersion);
+            $return->add(createPhpunitProcess($loopPhpVersion, $loopSymfonyVersion));
         }
     }
 
@@ -48,14 +48,14 @@ function createPhpunitProcess(string $phpVersion, string $symfonyVersion): Proce
 
 $phpVersion = null;
 $symfonyVersion = null;
-$applicationArgv = new StringArray();
+$applicationArgv = new StringCollection();
 foreach ($argv as $arg) {
     if (str_starts_with($arg, '--php=')) {
         $phpVersion = substr($arg, 6);
     } elseif (str_starts_with($arg, '--symfony=')) {
         $symfonyVersion = substr($arg, 10);
     } else {
-        $applicationArgv[] = $arg;
+        $applicationArgv->add($arg);
     }
 }