+++ src/Process/ProcessInterface.php
@@ -67,5 +67,7 @@ interface ProcessInterface
     /** @param array<mixed> $env */
     public function start(callable $callback = null, array $env = []): void;
 
+    public function stop(float $timeout = 10, int $signal = null): ?int;
+
     public function getStartTime(): float;
 }